refactoring, tests passed.
