+++ gqlauth/core/constants.py
