+++ tests/test_swap_emails.py
@@ -1,6 +1,6 @@
 import pytest
 
-from gqlauth.constants import Messages
+from gqlauth.core.constants import Messages
 
 from .testCases import (
     ArgTestCase,
