+++ tests/test_verify_account.py
@@ -1,6 +1,6 @@
-from gqlauth.constants import Messages
-from gqlauth.signals import user_verified
-from gqlauth.utils import get_token
+from gqlauth.core.constants import Messages
+from gqlauth.core.utils import get_token
+from gqlauth.user.signals import user_verified
 
 from .testCases import ArgTestCase, AsyncArgTestCase, AsyncRelayTestCase, RelayTestCase
